fix(RVCV-160): Class name 변경 및 테스트코드 Class 생성
