+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/reviewproperty/application/usecase/ReviewColumnUseCase.java
@@ -1,7 +1,9 @@
 package com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase;
 
 import com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase.request.UpdateLayoutRequest;
+import com.romanticpipe.reviewcanvas.reviewproperty.domain.ReviewColumn;
 
-public interface ReviewLayoutUseCase {
-	void updateLayout(Integer adminId, UpdateLayoutRequest updateLayoutRequest);
+public interface ReviewColumnUseCase {
+
+	ReviewColumn getColumn(Integer shopAdminId);
 }
