+++ api-module/src/test/java/com/romanticpipe/reviewcanvas/domain/reviewproperty/application/usecase/ReviewColumnUseCaseImplTest.java
@@ -1,0 +1,10 @@
+package com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase;
+
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+class ReviewColumnUseCaseImplTest {
+
+
+}
